+++ app/routes/components/ScrollArea/changelog.codegen.ts
@@ -3,4 +3,5 @@
 
 export const changelog = ''
   + '- 2025-10-04: Initial version.\n'
+  + '- 2025-10-05: Make the scrollbar appear on hover only.\n'
   ;